add new plotting algorithm
the new algorithm kicks in if there is more than one point of data
per pixel. It takes the first, last, minimum and maximum value from
the data points belonging to each pixel and adds them to the polygon.
That way, the displayed number of points is reduced to four times
the width of the plot window -- no matter the amount of data points --
without losing detail.